// DXC_Frontend/src/api/models/ocopProductCategoryDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface OcopProductCategoryDto {
  publicId?: string
  /** @nullable */
  name: string | null
  /** @nullable */
  description?: string | null
  /** @nullable */
  imageUrl?: string | null
  /** @nullable */
  imagePublicId?: string | null
  displayOrder?: number
  isActive?: boolean
  createdAt?: string
  /** @nullable */
  updatedAt?: string | null
}